Extract shared AbstractPostgresJournalPersistenceAdapter
Mirrors the digest A2 pattern: journal-pg and journal-pglite now share
one implementation of insertEntry, updateFullState, readEntry,
queryEntries, readFullState, queryFullStates, deleteByChart,
getCurrentTime, emitJournalEntryNotification and the row parsers,
written once against $N positional SQL in journal-persistence.

journal-pg keeps its 4-connection topology; the base routes to it via
three role-based hooks (runWriteQuery/runUpdateQuery/runReadQuery)
instead of a single query runner. The pg adapter's pg-bind bind(...)
calls and escapeLiteral string interpolation are replaced by the
parameterized VALUES-list approach already used in journal-pglite, and
pg-bind is dropped from journal-pg's dependencies.

Adds a shared decodeBytea helper to xjog-util so bytea columns are
decoded consistently across both drivers (pg's Buffer and PGlite's
Uint8Array), fixing the same String()-coercion drift previously fixed
for digests. Also fixes a latent bug where journal-pg's full-state SELECT
omitted the ownerId column.
