+++ packages/journal-persistence/src/index.ts
@@ -1,5 +1,9 @@
+export * from './AbstractPostgresJournalPersistenceAdapter';
 export * from './FullStateEntry';
 export * from './FullStateQuery';
 export * from './JournalEntry';
 export * from './JournalPersistenceAdapter';
 export * from './JournalQuery';
+export * from './JournalQueryRunner';
+export * from './PostgresFullStateRow';
+export * from './PostgresJournalRow';
