+++ packages/journal-pg/src/index.ts
@@ -1,3 +1,1 @@
-export * from './PostgresFullStateRow';
 export * from './PostgresJournalPersistenceAdapter';
-export * from './PostgresJournalRow';
