+++ packages/journal-pglite/src/index.ts
@@ -1,3 +1,1 @@
-export * from './PGliteFullStateRow';
 export * from './PGliteJournalPersistenceAdapter';
-export * from './PGliteJournalRow';
